+++ db/src/main/java/com/syntaric/openfhir/db/repository/mongodb/FhirConnectModelMongoRepository.java
@@ -21,4 +21,7 @@ public interface FhirConnectModelMongoRepository extends FhirConnectModelReposit
 
     @Query(value = "{'organisation': ?0}", delete = true)
     void deleteAllTenant(@NonNull String organisation);
+
+    @Query(value = "{'id': ?0, 'organisation': ?1}", delete = true)
+    void deleteByIdAndOrganisation(@NonNull String id, @NonNull String organisation);
 }